service/pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.136.0",
    "opentelemetry-distro[otlp]>=0.62b0",
    "opentelemetry-instrumentation>=0.62b0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.62b0",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "sqlalchemy>=2.0.49",
    "structlog>=25.5.0",
]
